backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.25.2
plotly==5.17.0
scikit-learn==1.3.2
statsmodels==0.14.0
dspy-ai==2.4.9
openai==1.3.7
anthropic==0.40.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
httpx==0.25.2
aiofiles==23.2.1
openpyxl==3.1.2
xlrd==2.0.1
matplotlib==3.8.2
seaborn==0.13.0
jupyter==1.0.0
ipykernel==6.26.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
pytz==2023.3
typing-extensions==4.8.0
